requirements.txt
openai>=1.0.0
jinja2
python-dotenv
requests
neo4j>=5.0.0
streamlit
pandas


